+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * [UIPQB-261](https://folio-org.atlassian.net/browse/UIPQB-261) Add maxColumnWidth support
 * [UIPQB-267](https://folio-org.atlassian.net/browse/UIPQB-267) *BREAKING* Use GET /locale to get tenant language & locale settings
 * [UIPQB-268](https://folio-org.atlassian.net/browse/UIPQB-268) Restore 'Select value' placeholder in value dropdown
+* [UIPQB-271](https://folio-org.atlassian.net/browse/UIPQB-271) Disambiguate duplicate language labels in UI
 
 ## [2.0.3](https://github.com/folio-org/ui-plugin-query-builder/tree/v2.0.3) (2025-04-18)